requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
python-dotenv==1.0.0
itsdangerous==2.1.2
sqlalchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7
alembic==1.13.1
cloud-sql-python-connector==1.11.0
